projectsample1/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
colorama==0.4.4
distlib==0.3.1
Django==3.2.3
django-cors-headers==3.7.0
django-extensions==3.1.3
django-phonenumber-field==5.2.0

djangorestframework==3.12.2
filelock==3.0.12
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
phonenumbers==8.12.24
Pillow==8.2.0
pycodestyle==2.7.0
PyJWT==1.7.1
pylint==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.6
wrapt==1.12.1